+++ src/Billpayer.php
@@ -12,8 +12,7 @@
 namespace Vanilo\Contracts;
 
 /**
- * A customer (either company or individual person) that
- * receives an Invoice
+ * A customer (either an organization or an individual person) that receives an Invoice
  */
 interface Billpayer extends Customer
 {